Add RSS controller
